Upgrade to rewrite-cljc test
This project uses both rewrite-clj and cljfmt which also use rewrite-clj.

I have upgraded to rewrite-cljc and my rewrite-cljc version of cljfmt.

Because rewrite-cljc is still in test and not on clojars yet,
I included it and the rewrite-cljc version of cljfmt via lein-git-down.

Notes of interest
- The upgrade of Clojure source was trivial on this project.
- This project supports Clojure 8 and above while rewrite-cljc is Clojure 9
and above. Ran tests with Clojure 1.101.536 and openjdk 1.8.0_252.
- I wonder if the paredit API brought into rewrite-cljc by rewrite-cljs would
be of any use in this project

1) Compile refactor-nrepl Java
lein javac

2) Run nrefactor-nrepl tests
lein test

3) Run tests on mranderson-ified code
make test